Kick Clips Support
Hello I Am From The Gta Rp Community And Streamers Like 4head Silent Ramee Are On Kick. So I Seen People Wish Kick Clips Work So I Have Added Support For Them So Streamer On Twitch Can React To Them. If There Anything I Need To Change For It To Be Able To Add It Lmk